check for main loop

+++ shiny_app/modules/light_speed.py
@@ -55,8 +55,9 @@ def get_website_prices(browser: webdriver.Safari, url: str):
     return json.loads(json_price)
 
 
-os.system("killall -u cbusillo 'Google Chrome'")
-driver = Driver(headless2=True, uc=True)
+if os.environ.get("RUN_MAIN", None) != "true":
+    os.system("killall -u cbusillo 'Google Chrome'")
+    driver = Driver(headless2=True, uc=True)
 
 
 class js_function_available:
